stratergy bug fix

+++ src/app/modules/core/strategy/containers/strategy-update/strategy-update.component.ts
@@ -114,7 +114,7 @@ export class StrategyUpdateComponent implements OnInit {
       }
       strategyJson[key] = formValues[key];
       if (this.strategyFormService.checkKeyForNumber(key)) {
-        strategyJson[key] = parseInt(formValues[key]);
+        strategyJson[key] = parseFloat(formValues[key]);
       }
       if (key === 'indicators' && formValues['direction'] === 'SELL') {
         strategyJson[key] = formValues[key]?.map((indi) => {
